docs: update vscode configuration template
It's unlikely for a project with a good amount of code and a transpilation step (i.e. ts + webpack) that the point where the debugger can attach is beyond the vscode default timeout.  To get it working I had to add this.


Additionally passing "foo" and "bar" by default configuration, although seemingly obvious, is just an unnecessary thing to have to delete.  I think either it's removed or keep it empty like above.   Also I'm not totally sure but I believe the standard debugging file is now `launch.json` and allows multiple configurations of which this can be one of.